Extract remote.js
