+++ lib/replication/staticreplication.js
@@ -1,12 +1,12 @@
 // USAGE
 // node lib/staticreplication.js /Users/user/Work/perserver/sling/staticreplication http://admin:admin@localhost:8080/content/tenant
 
-require('./string')
-
 const fs = require('fs')
 const http = require('http')
 const https = require('https')
-const axios = require('axios')
+
+require('./string')
+const remote = require('./remote')
 
 const RENDITIONS = '_/renditions'
 const THUMBNAIL = '.thumbnail.png'
@@ -14,7 +14,6 @@ const HTML = '.html'
 const INFINITY = '.infinity.json'
 const DATA = '.data.json'
 const SUFFIXES = [HTML, INFINITY, DATA]
-const NODES_INFO_PATH = '/perapi/admin/nodes.json'
 
 const args = process.argv.slice(2)
 let index = 0
@@ -173,56 +172,9 @@ function trimSuffix(path) {
 }
 
 function checkConsistencyFromStagePOV() {
-    grabNodesInfo(tenantPath)
-        .then(grabChildrenNodesInfo)
-        .then(extractRenderableResourcesFromNodesTree)
+    remote.grabFullNodesInfo(host, tenantPath)
         .then(array => array.filter(x => !filesLive.check.includes(rootDirPath + x)))
         .then(printStagePOVConsistencyInfo)
-}
-
-function grabNodesInfo(path) {
-    return axios.get(host + NODES_INFO_PATH + path)
-        .then(res => res.data)
-        .then(res => findChildOfPath(res, path))
-}
-
-function findChildOfPath(node, path) {
-    let result = node
-    while (result?.path !== path && result?.hasChildren && result?.children) {
-        result = result.children.filter(x => path.startsWith(`${x.path}/`) || x.path === path).pop()
-    }
-
-    return result
-}
-
-async function grabChildrenNodesInfo(node) {
-    const children = (node?.hasChildren && node?.children) || []
-    const fullChildren = await Promise.all(children.map(child => grabNodesInfo(child.path)))
-    for (let i = 0; i < children.length; i++) {
-        const child = children[i]
-        child.children = fullChildren[i].children || []
-        await grabChildrenNodesInfo(child)
-    }
-
-    return node
-}
-
-function extractRenderableResourcesFromNodesTree(root) {
-    let result = []
-    const { path, resourceType, children } = root
-    if (resourceType === 'per:Asset') {
-        result.push(path)
-    } else if (resourceType === 'per:Page') {
-        result.push(`${path}${HTML}`)
-        result.push(`${path}${DATA}`)
-    } else if (resourceType === 'per:Object') {
-        result.push(`${path}${INFINITY}`)
-    }
-
-    children.map(extractRenderableResourcesFromNodesTree).forEach((e, i, a) => {
-        result = result.concat(e)
-    })
-    return result
 }
 
 function printStagePOVConsistencyInfo(missingFiles) {
